python/requirements.txt
aws-cdk-lib==2.66.1
constructs>=10.0.0,<11.0.0
attini-cdk-lib>=1.2.1
